@@ -1,14 +1,12 @@
 module github.com/dgrr/http2
 
-go 1.24.0
+go 1.25.0
 
 require (
 	github.com/stretchr/testify v1.11.1
 	github.com/summerwind/h2spec v2.2.1+incompatible
 	github.com/valyala/fasthttp v1.68.0
 	github.com/valyala/fastrand v1.1.0
-	golang.org/x/crypto v0.46.0
-	golang.org/x/net v0.48.0
 )
 
 require (
@@ -20,6 +18,7 @@ require (
 	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	golang.org/x/net v0.48.0 // indirect
 	golang.org/x/sys v0.39.0 // indirect
 	golang.org/x/text v0.32.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
