@@ -52,7 +52,7 @@ type Stream struct {
 }
 
 var streamPool = sync.Pool{
-	New: func() interface{} {
+	New: func() any {
 		return &Stream{}
 	},
 }
